requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest==7.4.0
model-bakery==1.12.0
PyYAML==6.0.1
phonenumbers==8.13.19
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.1.0
djangorestframework-yaml==2.0.0
gunicorn==21.2.0
flake8==6.1.0